Add dialog utils separately
