+++ main.js
@@ -1,45 +1,12 @@
-const { app, BrowserWindow, ipcMain, dialog } = require('electron');
+const { app, BrowserWindow, ipcMain } = require('electron');
 const { is } = require('electron-util');
 const path = require('path');
-const fs = require('fs');
-const Store = require('electron-store');
+const { clientExists } = require( path.join(process.resourcePath, 'utils', 'client-check') );
 
-if (require('electron-squirrel-startup')) return app.quit();
 
-const store = new Store();
+if (require('electron-squirrel-startup')) return app.exit(1);
+
 let mainWindow;
-
-function selectClientPath() {
-    dialog
-        .showOpenDialog(mainWindow, {
-            title: 'Choose client folder',
-            properties: ['openDirectory'],
-        })
-        .then(({canceled, filePaths}) => {
-            if (filePaths.length > 0) {
-                const [path] = filePaths;
-                if (fs.existsSync(`${path}/WoWClient.exe`)) {
-                    store.set('clientPath', path);
-                } else {
-                    selectClientPath();
-                }
-            } else {
-
-            }
-
-            if (canceled) {
-                app.quit();
-            }
-        })
-        .catch((err) => {
-            dialog.showMessageBox({
-                title: 'Error while selecting the folder',
-                type: 'error',
-                message: 'The folder is either not readable or something has gone wrong, please raise an issue on github with the following details',
-                detail: JSON.stringify(err),
-            }, () => app.quit());
-        });
-}
 
 function createWindow() {
     mainWindow = new BrowserWindow({
@@ -55,30 +22,8 @@ function createWindow() {
         }
     });
 
-    const clientPath = store.get('clientPath') || '.';
-    if (!fs.existsSync(`${clientPath}/WoWClient.exe`)) {
-        dialog
-            .showMessageBox(mainWindow, {
-                title: 'WoWClient.exe not found',
-                type: 'warning',
-                detail: 'WoWClient executable not found, please move launcher to game folder or select the client folder.',
-                buttons: ['Exit', 'Choose client folder'],
-            })
-            .then(({response}) => {
-                switch (response) {
-                    case 1: {
-                        selectClientPath();
-                        return;
-                    }
-                    default: {
-                        app.quit();
-                        return;
-                    }
-                }
-                
-            });
-
-    }
+    // Ask for client path
+    clientExists();
 
     mainWindow.loadFile('src/index.html');
 }
