+++ package.json
@@ -46,6 +46,15 @@
         "directories": {
             "buildResources": "build",
             "output": "electron/output"
-        }
+        },
+        "extraResources": [
+            {
+                "from": "./src/utils/",
+                "to": "utils",
+                "filter": [
+                    "**/*"
+                ]
+            }
+        ]
     }
 }
